Containerize all services. Modify scripts.
