+++ DummyStockServer/Dockerfile
@@ -1,0 +1,5 @@
+FROM python:3.9.1-buster
+
+WORKDIR /usr/src/app
+COPY ./* ./
+CMD ["python", "./DummyStock.py"]
