+++ DummyStockServer/config.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+dummy_base_name=dummy-stock
+dummy_base_num=1
+
+dummy_service_port=4444
+
+dummy_service_image=$dummy_base_name:$dummy_base_num
+dummy_service_name=$dummy_base_name-$dummy_base_num
