+++ start_service.sh
@@ -1,5 +1,13 @@
-./RedisService/start_service.sh
-./DatabaseService/start_service.sh
-./AuditService/start_service.sh
-./TransactionService/start_service.sh
-./WebService/start_service.sh
+#!/bin/bash
+
+echo 'Creating docker network for containers...'
+sudo docker network create myNetwork >/dev/null 2>&1
+
+curr_path=$(dirname $0)
+
+$curr_path/DummyStockServer/start_service.sh
+$curr_path/RedisService/start_service.sh
+$curr_path/DatabaseService/start_service.sh
+$curr_path/AuditService/start_service.sh
+$curr_path/TransactionService/start_service.sh
+$curr_path/WebService/start_service.sh
